chore: rename rspack folder
