+++ packages/core/src/rspack-plugins/RsbuildHtmlPlugin.ts
