+++ packages/core/src/rspack-plugins/resource-hints/HtmlResourceHintsPlugin.ts
