+++ packages/core/src/rspack-plugins/resource-hints/extractChunks.ts
