+++ packages/core/src/rspack-plugins/resource-hints/getResourceType.ts
